+++ src/submission-service.ts
@@ -1,32 +1,19 @@
 import { isOffline } from './offline-service'
-import { getFormsKeyId, isLoggedIn } from './auth-service'
+import { isLoggedIn } from './auth-service'
 import {
   registerPendingQueueListener,
-  addFormSubmissionToPendingQueue,
   getPendingQueueSubmissions,
   getFormSubmission,
   updatePendingQueueSubmission,
   deletePendingQueueSubmission,
 } from './services/pending-queue'
-import {
-  checkForPaymentSubmissionEvent,
-  handlePaymentSubmissionEvent,
-} from './payment-service'
-import { generateSubmissionCredentials } from './services/api/submissions'
-import { uploadFormSubmission } from './services/s3Submit'
 import uploadAttachment from './services/uploadAttachment'
-import { deleteDraft } from './draft-service'
-import { removePrefillFormData } from './prefill-service'
+import { generateSubmissionCredentials } from './services/api/submissions'
 import replaceCustomValues from './services/replace-custom-values'
-import recentlySubmittedJobsService from './services/recently-submitted-jobs'
 import { FormTypes } from '@oneblink/types'
-import { getUserToken } from './services/user-token'
 import Sentry from './Sentry'
 import prepareSubmissionData from './services/prepareSubmissionData'
-import {
-  handleSchedulingSubmissionEvent,
-  checkForSchedulingSubmissionEvent,
-} from './services/schedulingHandlers'
+import submitForm, { SubmissionParams } from './services/submit'
 import {
   PendingFormSubmission,
   FormSubmission,
@@ -125,112 +112,11 @@ async function processPendingQueue(): Promise<void> {
   _isProcessingPendingQueue = false
 }
 
-async function submit({
-  formSubmission,
-  paymentReceiptUrl,
-  schedulingUrlConfiguration,
-}: {
-  formSubmission: FormSubmission
-  paymentReceiptUrl?: string
-  schedulingUrlConfiguration?: {
-    schedulingReceiptUrl: string
-    schedulingCancelUrl: string
-  }
-}): Promise<FormSubmissionResult> {
-  formSubmission.keyId = getFormsKeyId() || undefined
-  const paymentSubmissionEventConfiguration =
-    checkForPaymentSubmissionEvent(formSubmission)
-
-  const schedulingSubmissionEvent =
-    checkForSchedulingSubmissionEvent(formSubmission)
-
-  if (isOffline()) {
-    if (paymentSubmissionEventConfiguration || schedulingSubmissionEvent) {
-      console.log(
-        'Offline - form has a payment/scheduling submission event that has not been processed yet, return offline',
-        { paymentSubmissionEventConfiguration, schedulingSubmissionEvent },
-      )
-      return Object.assign({}, formSubmission, {
-        isOffline: true,
-        isInPendingQueue: false,
-        submissionTimestamp: null,
-        submissionId: null,
-        payment: null,
-        scheduling: null,
-      })
-    }
-
-    console.log('Offline - saving submission to pending queue..')
-    await addFormSubmissionToPendingQueue(formSubmission)
-    return Object.assign({}, formSubmission, {
-      isOffline: true,
-      isInPendingQueue: true,
-      submissionTimestamp: null,
-      submissionId: null,
-      payment: null,
-      scheduling: null,
-    })
-  }
-
-  const data = await generateSubmissionCredentials(formSubmission)
-
-  const formSubmissionResult: FormSubmissionResult = {
-    ...formSubmission,
-    payment: null,
-    scheduling: null,
-    isOffline: false,
-    isInPendingQueue: false,
-    submissionTimestamp: data.submissionTimestamp,
-    submissionId: data.submissionId,
-  }
-
-  if (schedulingSubmissionEvent && schedulingUrlConfiguration) {
-    formSubmissionResult.scheduling = await handleSchedulingSubmissionEvent({
-      formSubmissionResult,
-      schedulingSubmissionEvent,
-      schedulingUrlConfiguration,
-      paymentReceiptUrl,
-    })
-  } else if (paymentSubmissionEventConfiguration && paymentReceiptUrl) {
-    formSubmissionResult.payment = await handlePaymentSubmissionEvent({
-      ...paymentSubmissionEventConfiguration,
-      formSubmissionResult,
-      paymentReceiptUrl,
-    })
-  }
-  const userToken = getUserToken()
-
-  await uploadFormSubmission(
-    data,
-    {
-      formsAppId: formSubmission.formsAppId,
-      definition: formSubmission.definition,
-      submission: formSubmission.submission,
-      submissionTimestamp: data.submissionTimestamp,
-      keyId: formSubmission.keyId,
-      ipAddress: data.ipAddress,
-      user: data.userProfile,
-    },
-    {
-      externalId: formSubmission.externalId || undefined,
-      jobId: formSubmission.jobId || undefined,
-      userToken: userToken || undefined,
-      usernameToken: data.usernameToken,
-      previousFormSubmissionApprovalId:
-        formSubmissionResult.previousFormSubmissionApprovalId,
-    },
-  )
-  if (formSubmission.draftId) {
-    await deleteDraft(formSubmission.draftId, formSubmission.formsAppId)
-  }
-  if (formSubmission.preFillFormDataId) {
-    await removePrefillFormData(formSubmission.preFillFormDataId)
-  }
-  if (formSubmission.jobId) {
-    await recentlySubmittedJobsService.add(formSubmission.jobId)
-  }
-
-  return formSubmissionResult
+async function submit(params: SubmissionParams): Promise<FormSubmissionResult> {
+  return submitForm({
+    ...params,
+    generateCredentials: generateSubmissionCredentials,
+  })
 }
 
 async function closeWindow(): Promise<void> {